automatically backed up by learn

+++ index.js
@@ -1,1 +1,1 @@
-console.log(parseInt('2.2222', 10))
+console.log(parseFloat('2.2222', 10))
